BUGFIX and replace degree with radian.

+++ Mt3d.h
@@ -21,8 +21,8 @@ struct Mt3d
     int const width;
     int const height;
     
-    int /*const*/ alpha; // Horizontal range of view (degrees).
-    double /*const*/ beta; // Vertical range of view (degrees).
+    double /*const*/ alpha; // Horizontal range of view (radian).
+    double /*const*/ beta; // Vertical range of view (radian).
     
     double /*const*/ h; // Height of players eye as fraction of ceiling height (ceiling height is measured in cell lengths).
     
@@ -34,7 +34,7 @@ struct Mt3d
     double /*const*/ * const e; // One e value for each (x,y) pixel coordinate (in cell lengths).
     int /*const*/ * const floorY; // First y value that has "an e" reaching the floor and not the ceiling for each x.
     
-    double /*const*/ * const eta; // One value for each (x/y) pixel coordinate (in degrees).
+    double /*const*/ * const eta; // One value for each (x/y) pixel coordinate (in degree).
     
     // **********************
     // *** CURRENT VALUES ***
@@ -45,7 +45,7 @@ struct Mt3d
     double posX;
     double posY;
     
-    double gamma; // Angle (degrees) telling the player's view direction.
+    double gamma; // Angle (radian) telling the player's view direction.
     
     struct Map * map; // Does NOT take ownership.
     unsigned char * pixels; // Does NOT take ownership.
@@ -53,8 +53,8 @@ struct Mt3d
 
 void Mt3d_draw(struct Mt3d * const inObj);
 void Mt3d_delete(struct Mt3d * const inObj);
-void Mt3d_update(int const inAlpha, double const inBeta, double const inH, struct Mt3d * const inOutObj);
-struct Mt3d * Mt3d_create(int const inWidth, int const inHeight, int const inAlpha, double const inBeta, double const inH);
+void Mt3d_update(double const inAlpha, double const inBeta, double const inH, struct Mt3d * const inOutObj);
+struct Mt3d * Mt3d_create(int const inWidth, int const inHeight, double const inAlpha, double const inBeta, double const inH);
 
 #ifdef __cplusplus
 }
